Divorce the pbft and statetransfer mock testing infrastructure
