+++ consensus/obcpbft/mock_consumer_test.go
@@ -108,8 +108,7 @@ func makeConsumerNetwork(N int, makeConsumer func(id uint64, config *viper.Viper
 			testEndpoint: tep,
 		}
 
-		ml := NewMockLedger(&twl, nil)
-		ml.PutBlock(0, SimpleGetBlock(0)) // Initialize a genesis block
+		ml := NewMockLedger(&twl)
 		twl.mockLedgers[id] = ml
 
 		cs := &completeStack{
