+++ consensus/statetransfer/statetransfer_mock_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package obcpbft
+package statetransfer
 
 import (
 	"bytes"
@@ -139,8 +139,6 @@ type MockLedger struct {
 
 	deltaID       interface{}
 	preDeltaValue uint64
-
-	ce *consumerEndpoint // To support the ExecTx stuff
 }
 
 func NewMockLedger(remoteLedgers LedgerDirectory, filter func(request mockRequest, peerID *protos.PeerID) mockResponse) *MockLedger {
@@ -161,123 +159,6 @@ func NewMockLedger(remoteLedgers LedgerDirectory, filter func(request mockReques
 	mock.remoteLedgers = remoteLedgers
 
 	return mock
-}
-
-func (mock *MockLedger) BeginTxBatch(id interface{}) error {
-	if mock.txID != nil {
-		return fmt.Errorf("Tx batch is already active")
-	}
-	mock.txID = id
-	mock.curBatch = nil
-	mock.curResults = nil
-	mock.preBatchState = mock.state
-	return nil
-}
-
-func (mock *MockLedger) ExecTxs(id interface{}, txs []*protos.Transaction) ([]byte, error) {
-	if !reflect.DeepEqual(mock.txID, id) {
-		return nil, fmt.Errorf("Invalid batch ID")
-	}
-
-	mock.curBatch = append(mock.curBatch, txs...)
-	var err error
-	var txResult []byte
-	if nil != mock.ce && nil != mock.ce.execTxResult {
-		txResult, err = mock.ce.execTxResult(txs)
-	} else {
-		// This is basically a default fake default transaction execution
-		if nil == txs {
-			txs = []*protos.Transaction{&protos.Transaction{Payload: SimpleGetStateDelta(mock.blockHeight)}}
-		}
-
-		for _, transaction := range txs {
-			if transaction.Payload == nil {
-				transaction.Payload = SimpleGetStateDelta(mock.blockHeight)
-			}
-
-			txResult = append(txResult, transaction.Payload...)
-		}
-
-	}
-
-	buffer := make([]byte, binary.MaxVarintLen64)
-
-	for i, b := range txResult {
-		buffer[i%binary.MaxVarintLen64] += b
-	}
-
-	mock.ApplyStateDelta(id, SimpleBytesToStateDelta(buffer))
-
-	mock.curResults = append(mock.curResults, txResult...)
-
-	return txResult, err
-}
-
-func (mock *MockLedger) CommitTxBatch(id interface{}, metadata []byte) (*protos.Block, error) {
-	block, err := mock.commonCommitTx(id, metadata, false)
-	if nil == err {
-		mock.txID = nil
-		mock.curBatch = nil
-		mock.curResults = nil
-	}
-	return block, err
-}
-
-func (mock *MockLedger) commonCommitTx(id interface{}, metadata []byte, preview bool) (*protos.Block, error) {
-	if !reflect.DeepEqual(mock.txID, id) {
-		return nil, fmt.Errorf("Invalid batch ID")
-	}
-
-	previousBlockHash := []byte("Genesis")
-	if 0 < mock.blockHeight {
-		previousBlock, _ := mock.GetBlock(mock.blockHeight - 1)
-		previousBlockHash, _ = mock.HashBlock(previousBlock)
-	}
-
-	stateHash, _ := mock.GetCurrentStateHash()
-
-	block := &protos.Block{
-		ConsensusMetadata: metadata,
-		PreviousBlockHash: previousBlockHash,
-		StateHash:         stateHash,
-		Transactions:      mock.curBatch,
-		NonHashData: &protos.NonHashData{
-			TransactionResults: []*protos.TransactionResult{
-				&protos.TransactionResult{
-					Result: mock.curResults,
-				},
-			},
-		},
-	}
-
-	if !preview {
-		if nil != mock.CommitStateDelta(id) {
-			panic("Error in delta construction/application")
-		}
-		fmt.Printf("TEST LEDGER: Mock ledger is inserting block %d with hash %x\n", mock.blockHeight, SimpleHashBlock(block))
-		mock.PutBlock(mock.blockHeight, block)
-	}
-
-	return block, nil
-}
-
-func (mock *MockLedger) PreviewCommitTxBatch(id interface{}, metadata []byte) ([]byte, error) {
-	b, err := mock.commonCommitTx(id, metadata, true)
-	if err != nil {
-		return nil, err
-	}
-	return mock.getBlockInfoBlob(mock.blockHeight+1, b), nil
-}
-
-func (mock *MockLedger) RollbackTxBatch(id interface{}) error {
-	if !reflect.DeepEqual(mock.txID, id) {
-		return fmt.Errorf("Invalid batch ID")
-	}
-	mock.curBatch = nil
-	mock.curResults = nil
-	mock.txID = nil
-	mock.state = mock.preBatchState
-	return nil
 }
 
 func (mock *MockLedger) GetBlockchainSize() uint64 {
@@ -657,37 +538,6 @@ func (mock *MockLedger) VerifyBlockchain(start, finish uint64) (uint64, error) {
 		}
 
 		current = next
-	}
-}
-
-func (mock *MockLedger) GetBlockchainInfoBlob() []byte {
-	b, _ := mock.GetBlock(mock.blockHeight - 1)
-	return mock.getBlockInfoBlob(mock.blockHeight, b)
-}
-
-func (mock *MockLedger) getBlockInfoBlob(height uint64, block *protos.Block) []byte {
-	info := &protos.BlockchainInfo{Height: height}
-	info.CurrentBlockHash, _ = mock.HashBlock(block)
-	h, _ := proto.Marshal(info)
-	return h
-}
-
-func (mock *MockLedger) GetBlockHeadMetadata() ([]byte, error) {
-	b, ok := mock.blocks[mock.blockHeight-1]
-	if !ok {
-		return nil, fmt.Errorf("could not retrieve block from mock ledger")
-	}
-	return b.ConsensusMetadata, nil
-}
-
-func (mock *MockLedger) simulateStateTransfer(meta []byte, id []byte, peers []*protos.PeerID) {
-	info := &protos.BlockchainInfo{}
-	proto.Unmarshal(id, info)
-	fmt.Printf("TEST LEDGER skipping to %+v, %+v", meta, info)
-	for n := mock.blockHeight; n < info.Height; n++ {
-		block := SimpleGetBlock(n)
-		block.ConsensusMetadata = meta
-		mock.PutBlock(n, block)
 	}
 }
